Drop OOM-killer score adjustment for Linux
Delete the code to adjust the OOM-killer score on Linux. OOM-killer score
adjustment is best performed by a separate script or utility.

Delete the `chmod u+s` command (to set the setuid permission bit on the
binary) from the Makefile install target.

Since commit c67ad670bc07b2b4e3852ba45775ee47b7f28b27, which replaced
support for crypt and BSD auth with PAM support, slock-pam is not intended
to be run as a privileged user, and it does not drop privileges if it is.