Fix bad clearing of password buffer
Do not try to clear the entered password buffer in readpw; the buffer needs
to be left intact after readpw and pam_conv return in order for the
conversation mechanism to read it.

Correctly clear the password buffer, rather than a reference to it, when
escape is pressed.

This commit fixes commit 51d3727d57d2225a207adec95ba12894f6006654.